[refactor] Rearrange file structure
