+++ kWordle/View/Shared/MailView.swift
@@ -9,6 +9,8 @@ import SwiftUI
 import UIKit
 import MessageUI
 
+//Not used for now
+
 struct ComposeMailData {
     let subject: String
     let recipients: [String]?
